Add Dependencia Hateoas e configuração de Link

+++ src/main/java/br/com/sysgestao/sysgestao/service/ClienteService.java
@@ -2,11 +2,11 @@ package br.com.sysgestao.sysgestao.service;
 
 import java.util.List;
 
+import br.com.sysgestao.sysgestao.domain.Cliente;
+import br.com.sysgestao.sysgestao.error.NotFoundException;
+import br.com.sysgestao.sysgestao.repository.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import br.com.sysgestao.sysgestao.domain.Cliente;
-import br.com.sysgestao.sysgestao.repository.ClienteRepository;
 
 @Service
 public class ClienteService {
@@ -19,18 +19,29 @@ public class ClienteService {
     }
 
     public Cliente findById(Long id) {
-        return clienteRepository.findById(id).get();
+        return clienteRepository.findById(id).orElseThrow(() -> new NotFoundException("Cliente não encontrado. id=" + id));
     }
 
     public void save(Cliente cliente) {
         clienteRepository.save(cliente);
     }
 
-    public Cliente update(Cliente cliente) {
-        return clienteRepository.saveAndFlush(cliente);
+    public void update(Cliente cliente) {
+        if (!clienteRepository.existsById(cliente.getId())) {
+            throw new NotFoundException("Cliente não encontrado: id=" + cliente.getId());
+        }
+
+        clienteRepository.saveAndFlush(cliente);
     }
 
-    public void deleteClientById(Long id) {
+    public void deletePorId(Long id) {
+        if (!clienteRepository.existsById(id)) {
+            throw new NotFoundException("Cliente não encontrado. id=" + id);
+        }
+
         clienteRepository.deleteById(id);
     }
+
+
 }
+
